+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -608,8 +608,6 @@ Other Operations
 
 .. autofunction:: tt_lib.tensor.repeat
 
-.. autofunction:: tt_lib.tensor.repeat_interleave
-
 .. autofunction:: tt_lib.tensor.pow
 
 .. autofunction:: tt_lib.tensor.argmax
