+++ tests/ttnn/unit_tests/gtests/test_repeat_interleave.cpp
@@ -8,7 +8,7 @@
 #include "ttnn/device.hpp"
 #include "ttnn/operations/core/core.hpp"
 #include "ttnn/async_runtime.hpp"
-#include "ttnn/operations/data_movement.hpp"
+#include "ttnn/operations/data_movement/repeat_interleave/repeat_interleave.hpp"
 #include "tt_numpy/functions.hpp"
 #include "tt_metal/common/logger.hpp"
