+++ ttnn/cpp/ttnn/operations/data_movement.hpp
@@ -85,28 +85,11 @@ struct Repeat {
     }
 };
 
-struct RepeatInterleave {
-
-    // # This operation does not support the following cases:
-    // #   - Shape([2[32], 2[32]]) -> repeats = 2, dim = 0
-    // #   - Shape([2[32], 2[32]]) -> repeats = Tensor[1,2], dim = 1
-    static ttnn::Tensor operator()(
-        const ttnn::Tensor& input_tensor,
-        uint32_t repeats,
-        int32_t dim,
-        std::optional<MemoryConfig> output_mem_config = std::nullopt) {
-        MemoryConfig mem_config = output_mem_config.value_or(input_tensor.memory_config());
-        auto output_tensor = tt::tt_metal::repeat_interleave(input_tensor, repeats, dim, mem_config);
-        return output_tensor;
-    }
-};
 
 }  // namespace data_movement
 }  // namespace operations
 
 constexpr auto upsample = ttnn::register_operation_with_auto_launch_op<"ttnn::upsample", ttnn::operations::data_movement::UpSample>();
 constexpr auto repeat = ttnn::register_operation_with_auto_launch_op<"ttnn::repeat", ttnn::operations::data_movement::Repeat>();
-constexpr auto repeat_interleave =
-    ttnn::register_operation_with_auto_launch_op<"ttnn::repeat_interleave", ttnn::operations::data_movement::RepeatInterleave>();
 
 }  // namespace ttnn
