+++ ttnn/cpp/ttnn/operations/data_movement/data_movement_pybind.hpp
@@ -16,6 +16,7 @@
 #include "ttnn/operations/data_movement/slice/slice_pybind.hpp"
 #include "ttnn/operations/data_movement/tilize/tilize_pybind.hpp"
 #include "ttnn/operations/data_movement/tilize_with_val_padding/tilize_with_val_padding_pybind.hpp"
+#include "ttnn/operations/data_movement/repeat_interleave/repeat_interleave_pybind.hpp"
 
 namespace py = pybind11;
 
@@ -41,43 +42,6 @@ void bind_upsample(py::module& module) {
         doc,
         ttnn::pybind_arguments_t{
             py::arg("input_tensor"), py::arg("scale_factor"), py::arg("memory_config") = std::nullopt});
-}
-
-void bind_repeat_interleave(py::module& module) {
-    auto doc = R"doc(
-repeat_interleave(input_tensor: ttnn.Tensor, repeats : int, dim: int = 0) -> ttnn.Tensor
-
-Repeats elements of a :attr:`tensor` in the given :attr:`dim`.
-
-Args:
-    * :attr:`input_tensor`: the input_tensor to apply the repeate interleave operation.
-    * :attr:`repeats`: The number of repetitions for each element. repeats is broadcasted to fit the shape of the given axis.
-    * :attr:`dim`: the dimension to expand with the repetitions.
-
-Example:
-
-torch_input_tensor =
-    torch_result = torch.repeat_interleave(torch_input_tensor, repeats, dim=dim)
-
-    input_tensor = ttnn.from_torch(torch_input_tensor, layout=ttnn.TILE_LAYOUT, device=device)
-
-    output = ttnn.repeat_interleave(input_tensor, repeats, dim=dim)
-    >>> a = ttnn.from_torch(torch.rand(1, 1, 32, 32, dtype=torch.bfloat16), layout=ttnn.TILE_LAYOUT, device=device)
-    >>> b = ttnn.repeat_interleave(a, 2, dim=0)
-    >>> print(a.shape, b.shape)
-    ttnn.Shape([1, 1, 32, 32]) ttnn.Shape([2, 1, 32, 32])
-        )doc";
-
-    ttnn::bind_registered_operation(
-        module,
-        ttnn::repeat_interleave,
-        doc,
-        ttnn::pybind_arguments_t{
-            py::arg("input_tensor"),
-            py::arg("repeats"),
-            py::arg("dim"),
-            py::kw_only(),
-            py::arg("memory_config") = std::nullopt});
 }
 
 void bind_repeat(py::module& module) {
@@ -119,7 +83,7 @@ void py_module(py::module& module) {
     detail::bind_slice(module);
     detail::bind_downsample(module);
     bind_repeat(module);
-    bind_repeat_interleave(module);
+    detail::bind_repeat_interleave(module);
     detail::bind_tilize(module);
     detail::bind_tilize_with_val_padding(module);
     detail::bind_tilize_with_zero_padding(module);
